Добавлена папка pages, на каждой странице добавлена кнопка возвращения назад
