+++ prog5nav/lib/pages/list1_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 
-class List2Screen extends StatelessWidget {
+class List1Screen extends StatelessWidget {
 
   // This widget is the root of your application.
   @override
@@ -12,11 +12,21 @@ class List2Screen extends StatelessWidget {
         appBar: AppBar(
           title: Text('Список элементов'),
           backgroundColor: Colors.green,
+          leading: IconButton(
+            icon: Icon(Icons.arrow_back),
+            onPressed: (){
+              Navigator.pop(context);
+            },
+          ),
         ),
-        body: ListView.builder(
-            itemBuilder: (context, index) {
-              return Text('Строка $index');
-            }
+        body: ListView(
+          children: const [
+            Text('0000'),
+            Divider(),
+            Text('0001'),
+            Divider(),
+            Text('0010')
+          ],
         ),
       ),
     );
