+++ prog5nav/lib/pages/list2_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 
-class List1Screen extends StatelessWidget {
+class List2Screen extends StatelessWidget {
 
   // This widget is the root of your application.
   @override
@@ -12,15 +12,17 @@ class List1Screen extends StatelessWidget {
         appBar: AppBar(
           title: Text('Список элементов'),
           backgroundColor: Colors.green,
+          leading: IconButton(
+            icon: Icon(Icons.arrow_back),
+            onPressed: (){
+              Navigator.pop(context);
+            },
+          ),
         ),
-        body: ListView(
-          children: const [
-            Text('0000'),
-            Divider(),
-            Text('0001'),
-            Divider(),
-            Text('0010')
-          ],
+        body: ListView.builder(
+            itemBuilder: (context, index) {
+              return Text('Строка $index');
+            }
         ),
       ),
     );
